refactor(utils): Reorganize the thumbnail utilities
- Moved all thumbnail utilities to `utils/thumb-utils` module
- All thumbnail utility functions are now within a namespace called `ThumbnailUtils` (the alias is `ThumbUtils`).

This change is intended for better module organization and maintainability for feature changes.
